fix to working version

+++ upsmonit/ups.py
@@ -2,34 +2,30 @@
 # -*- coding: utf-8 -*-
 
 import argparse
-import select
-import threading
-
 import asyncio
-from ph4runner import AsyncRunner, install_sarge_filter
-
-import coloredlogs
-import logging
 import json
-import socket
-import itertools
-import shlex
-import time
-import queue
-import sys
+import logging
 import os
-import random
-import hashlib
-from jsonpath_ng import parse
-from typing import Optional, List
+import platform
+import queue
+import select
+import signal
+import socket
+import socketserver
+import threading
+import time
 from queue import Queue
 
+import coloredlogs
+import jwt
+from jsonpath_ng import parse
+from ph4runner import AsyncRunner, install_sarge_filter
 from telegram import Update, User
 from telegram.error import TelegramError
 from telegram.ext import ApplicationBuilder, ContextTypes, CommandHandler
 
 logger = logging.getLogger(__name__)
-coloredlogs.install(level=logging.DEBUG)
+coloredlogs.install(level=logging.INFO)
 
 
 def try_fnc(fnc):
@@ -102,21 +98,29 @@ class UpsMonit:
         self.args = {}
         self.ups_name = 'servers'
         self.config = {}
+        self.jwt_key = None
         self.bot_apikey = None
         self.server_fifo = None
+        self.server_port = None
+        self.server_host = '127.0.0.1'
         self.allowed_usernames = []
         self.allowed_userids = []
         self.registered_chat_ids = []
         self.registered_chat_ids_set = set()
-        self.task_queue = Queue()
+        self.use_server = True
+        self.use_fifo = False
 
+        self.task_queue = Queue()
         self.is_running = True
         self.status_thread = None
         self.status_thread_last_check = 0
         self.last_ups_status = None
         self.last_ups_status_time = 0
         self.fifo_thread = None
+        self.server_thread = None
+        self.server_tcp = None
         self.main_loop = None
+        self.start_error = None
 
         self.bot_app = None
         self.bot_thread = None
@@ -139,7 +143,7 @@ class UpsMonit:
                             help='Event from the nut daemon')
         parser.add_argument('-u', '--users', dest='users', nargs=argparse.ZERO_OR_MORE,
                             help='Allowed user names')
-        parser.add_argument('--user-ids', dest='user_ids', nargs=argparse.ZERO_OR_MORE,
+        parser.add_argument('--user-ids', dest='user_ids', nargs=argparse.ZERO_OR_MORE, type=int,
                             help='Allowed user IDs')
         parser.add_argument('-t', '--chat-id', dest='chat_ids', nargs=argparse.ZERO_OR_MORE, type=int,
                             help='Pre-Registered chat IDs')
@@ -151,6 +155,8 @@ class UpsMonit:
 
     def load_config(self):
         self.bot_apikey = os.getenv('BOT_APIKEY', None)
+        self.jwt_key = os.getenv('JWT_KEY', None)
+        self.ups_name = os.getenv('UPS_NAME', None) or self.args.ups_name
 
         if not self.args.config:
             return
@@ -160,11 +166,21 @@ class UpsMonit:
                 dt = fh.read()
                 self.config = json.loads(dt)
 
-            self.bot_apikey = jsonpath('$.bot_apikey', self.config, True)
+            bot_apikey = jsonpath('$.bot_apikey', self.config, True)
+            if not self.bot_apikey:
+                self.bot_apikey = bot_apikey
+
+            jwt_key = jsonpath('$.jwt_key', self.config, True) or 'default-jwt-key-0x043719de'
+            if not self.jwt_key:
+                self.jwt_key = jwt_key
 
             ups_name = jsonpath('$.ups_name', self.config, True)
             if not self.ups_name:
                 self.ups_name = ups_name
+
+            server_port = jsonpath('$.server_port', self.config, True)
+            if not self.server_port:
+                self.server_port = server_port
 
             server_fifo = jsonpath('$.server_fifo', self.config, True)
             if not self.server_fifo:
@@ -185,6 +201,16 @@ class UpsMonit:
         except Exception as e:
             logger.error("Could not load config %s at %s" % (e, self.args.config), exc_info=e)
 
+    def _stop_app_on_signal(self):
+        logger.info(f'Signal received')
+        self.is_running = False
+
+    def init_signals(self):
+        stop_signals = (signal.SIGINT, signal.SIGTERM, signal.SIGABRT) if platform.system() != "Windows" else []
+        loop = asyncio.get_event_loop()
+        for sig in stop_signals or []:
+            loop.add_signal_handler(sig, self._stop_app_on_signal)
+
     def init_bot(self):
         self.bot_app = ApplicationBuilder().token(self.bot_apikey).build()
         start_handler = CommandHandler('start', self.bot_cmd_start)
@@ -196,6 +222,8 @@ class UpsMonit:
         self.bot_app.add_handler(status_handler)
 
     def load_bot_thread(self):
+        """Running bot in a separate thread. Experimental method.
+        Message handling does not work"""
         if not self.bot_apikey:
             logger.info('Telegram bot API key not configured')
             return
@@ -259,20 +287,25 @@ class UpsMonit:
             logger.info('Telegram bot API key not configured')
             return
 
-        self.init_bot()
-
         def error_callback(exc: TelegramError) -> None:
             logger.info(f'Error callback {exc}')
             self.bot_app.create_task(self.bot_app.process_error(error=exc, update=None))
 
-        await self.bot_app.initialize()
-        if self.bot_app.post_init:
-            await self.bot_app.post_init(self.bot_app)
-        await self.bot_app.updater.start_polling(error_callback=error_callback)
+        try:
+            self.init_bot()
+            await self.bot_app.initialize()
+            if self.bot_app.post_init:
+                await self.bot_app.post_init(self.bot_app)
+            await self.bot_app.updater.start_polling(error_callback=error_callback)
 
-        logger.info('Bot app start')
-        await self.bot_app.start()
-        logger.info('Bot started')
+            logger.info('Bot app start')
+            await self.bot_app.start()
+            logger.info('Bot started')
+
+        except Exception as e:
+            logger.error(f'Error starting telegram bot {e}', exc_info=e)
+            self.start_error = e
+            raise
 
     async def stop_bot(self):
         if not self.bot_app:
@@ -322,6 +355,16 @@ class UpsMonit:
 
         def fifo_internal():
             logger.info('Starting fifo thread')
+            try:
+                self.destroy_fifo()
+                self.create_fifo()
+                with open(self.server_fifo) as fifo:
+                    pass
+            except Exception as e:
+                logger.error(f'Error starting server fifo: {e}', exc_info=e)
+                self.start_error = e
+                return
+
             with open(self.server_fifo) as fifo:
                 while self.is_running:
                     try:
@@ -343,6 +386,39 @@ class UpsMonit:
         self.fifo_thread.daemon = False
         self.fifo_thread.start()
 
+    def start_server(self):
+        monit = self
+
+        class TcpServerHandler(socketserver.BaseRequestHandler):
+            def handle(self):
+                try:
+                    data = self.request.recv(8192).strip()
+                    r = monit.process_server_data(data.decode())
+                    self.request.sendall(r)
+                except Exception as e:
+                    logger.warning(f'Exception processing server message {e}', exc_info=e)
+                    self.request.sendall(json.dumps({'response': 500}).encode())
+
+        def server_internal():
+            logger.info('Starting server thread')
+            try:
+                self.server_tcp = socketserver.TCPServer((self.server_host, self.server_port), TcpServerHandler)
+                self.server_tcp.allow_reuse_address = True
+                self.server_tcp.serve_forever()
+            except Exception as e:
+                self.start_error = e
+                logger.error(f'Error in starting server thread {e}', exc_info=e)
+            finally:
+                logger.info('Stopping server thread')
+
+        self.server_thread = threading.Thread(target=server_internal, args=())
+        self.server_thread.daemon = False
+        self.server_thread.start()
+
+    def stop_server(self):
+        if self.server_tcp:
+            self.server_tcp.shutdown()
+
     def create_fifo(self):
         if not self.server_fifo:
             return
@@ -352,9 +428,16 @@ class UpsMonit:
         if not self.server_fifo:
             return
         try:
-            os.unlink(self.server_fifo)
+            if os.path.exists(self.server_fifo):
+                os.unlink(self.server_fifo)
         except Exception as e:
-            logger.error(f'Error unlinking fifo {e}', exc_info=e)
+            logger.warning(f'Error unlinking fifo {e}', exc_info=e)
+
+    def create_jwt(self, payload):
+        return jwt.encode(payload=payload, key=self.jwt_key, algorithm="HS256")
+
+    def decode_jwt(self, payload):
+        return jwt.decode(payload, self.jwt_key, algorithms=["HS256"])
 
     def main(self):
         install_sarge_filter()
@@ -362,15 +445,18 @@ class UpsMonit:
 
         parser = self.argparser()
         self.args = parser.parse_args()
+        if self.args.debug:
+            coloredlogs.install(level=logging.DEBUG)
+
         self.ups_name = self.args.ups_name
         self.allowed_usernames = self.args.users or []
         self.allowed_userids = self.args.user_ids or []
         self.registered_chat_ids = self.args.chat_ids or []
         self.server_fifo = self.args.server_fifo
+        self.server_port = self.args.server_port
         self.main_loop = asyncio.get_event_loop()
 
         self.load_config()
-        self.start_status_thread()
         self.registered_chat_ids_set = set(self.registered_chat_ids)
 
         # Async switch
@@ -388,58 +474,72 @@ class UpsMonit:
 
         # UPS generated event, should send notification anyway
         if self.args.event:
-            r = await self.event_handler()
+            return await self.event_handler()
 
-        else:
-            self.destroy_fifo()
-            self.create_fifo()
-            self.start_fifo_thread()
-            # self.load_bot_thread()
+        # Normal daemon mode
+        try:
+            self.init_signals()
+            self.start_status_thread()
+            if self.use_fifo:
+                self.start_fifo_thread()
+            if self.use_server:
+                self.start_server()
             await self.load_bot_async()
 
-            r = await self.main_handler()
-            self.destroy_fifo()
+            if self.start_error:
+                logger.error(f'Cannot continue, start error: {self.start_error}')
+                raise self.start_error
 
-        await self.stop_bot()
+            r = await self.main_handler()
+
+        finally:
+            if self.use_fifo:
+                try_fnc(lambda: self.destroy_fifo())
+            if self.use_server:
+                try_fnc(lambda: self.stop_server())
+            await self.stop_bot()
+
         return r
 
     def is_user_allowed(self, user: User):
         if not user:
             return False
+
         if user.id in self.allowed_userids:
             return True
-        if user.name in self.allowed_usernames:
+
+        if user.username in self.allowed_usernames:
             return True
         return False
 
-    async def bot_cmd_start(self, update: Update, context: ContextTypes.DEFAULT_TYPE):
+    async def reject_user(self, update: Update, context: ContextTypes.DEFAULT_TYPE):
+        await context.bot.send_message(chat_id=update.effective_chat.id, text=f"Fuck off")
+
+    async def check_user(self, method, update: Update, context: ContextTypes.DEFAULT_TYPE):
         user_allowed = self.is_user_allowed(update.message.from_user)
-        logger.info(f'New start message with chat_id: {update.effective_chat.id}, from {update.message.from_user}, '
+        logger.info(f'New "{method}" message with chat_id: {update.effective_chat.id}, from {update.message.from_user}, '
                     f'allowed {user_allowed}')
         if not user_allowed:
             await context.bot.send_message(chat_id=update.effective_chat.id, text=f"Fuck off")
+            return False
+        return True
+
+    async def bot_cmd_start(self, update: Update, context: ContextTypes.DEFAULT_TYPE):
+        if not await self.check_user("start", update, context):
             return
 
         await context.bot.send_message(chat_id=update.effective_chat.id, text="Registered")
         self.registered_chat_ids_set.add(update.effective_chat.id)
 
     async def bot_cmd_stop(self, update: Update, context: ContextTypes.DEFAULT_TYPE):
-        user_allowed = self.is_user_allowed(update.message.from_user)
-        logger.info(f'New stop message with chat_id: {update.effective_chat.id}, from {update.message.from_user}, '
-                    f'allowed {user_allowed}')
-        if not user_allowed:
-            await context.bot.send_message(chat_id=update.effective_chat.id, text=f"Fuck off")
+        if not await self.check_user("stop", update, context):
             return
 
         await context.bot.send_message(chat_id=update.effective_chat.id, text="Deregistering you")
         self.registered_chat_ids_set.remove(update.effective_chat.id)
 
     async def bot_cmd_status(self, update: Update, context: ContextTypes.DEFAULT_TYPE):
-        user_allowed = self.is_user_allowed(update.message.from_user)
-        logger.info(f'New status message with chat_id: {update.effective_chat.id}, from {update.message.from_user}, '
-                    f'allowed {user_allowed}')
-        if not user_allowed:
-            await context.bot.send_message(chat_id=update.effective_chat.id, text=f"Fuck off")
+        if not await self.check_user("status", update, context):
             return
 
         if time.time() - self.last_cmd_status < 3:
@@ -457,13 +557,11 @@ class UpsMonit:
     async def event_handler(self):
         notif_type = os.getenv('NOTIFYTYPE')
         logger.info(f'Event OS: {os.environ}, notif type: {notif_type}')
+        payload = {'type': 'event', 'notif': notif_type}
 
-        with open(self.server_fifo, 'w') as f:
-            f.write(json.dumps({'type': 'event', 'notif': notif_type}) + '\n')
-            f.flush()
+        self.send_daemon_message(payload)
 
     async def main_handler(self):
-        # TODO: register signal handlers to gracefully shutdown, also take inspiration from telegram bot
         while self.is_running:
             try:
                 next_task = self.task_queue.get(False)
@@ -471,19 +569,55 @@ class UpsMonit:
                     await asyncio.sleep(0.01)
                 await next_task
             except queue.Empty:
-                await asyncio.sleep(0.01)
+                await asyncio.sleep(0.02)
+        logger.info(f'Main thread finishing')
 
     async def send_telegram_notif(self, notif):
         for chat_id in self.registered_chat_ids_set:
             logger.info(f'Sending telegram notif {notif}, chat id: {chat_id}')
             await self.bot_app.bot.send_message(chat_id, notif)
 
+    def send_daemon_message(self, payload):
+        if self.use_server:
+            self.send_server_msg(payload)
+        elif self.use_fifo:
+            self.send_fifo_msg(payload)
+        else:
+            raise Exception('No connection method to the daemon')
+
+    def send_fifo_msg(self, payload):
+        with open(self.server_fifo, 'w') as f:
+            token = self.create_jwt(payload)
+            f.write(token + '\n')
+            f.flush()
+
+    def send_server_msg(self, payload):
+        tcp_client = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
+        try:
+            data = self.create_jwt(payload)
+            tcp_client.connect((self.server_host, self.server_port))
+            tcp_client.sendall((data + '\n').encode())
+
+            # Read data from the TCP server and close the connection
+            received = tcp_client.recv(8192).decode()
+            return received
+        finally:
+            tcp_client.close()
+
     def process_fifo_data(self, data):
         logger.info(f'Data read from fifo: {data}, len: {len(data)}')
+        self.process_message(data)
+
+    def process_server_data(self, data):
+        logger.info(f'TCP server data received: {data}')
+        self.process_message(data)
+        return json.dumps({'response': 200}).encode()
+
+    def process_message(self, data):
         lines = data.splitlines(False)
         for line in lines:
             try:
-                js = json.loads(line)
+                js = self.decode_jwt(line)
                 if 'type' not in js:
                     continue
 
@@ -492,9 +626,6 @@ class UpsMonit:
                     notif = js['notif']
 
                     self.task_queue.put(self.send_telegram_notif(f'UPS event: {notif}'))
-                    # asyncio.run_coroutine_threadsafe(self.send_telegram_notif(f'UPS event: {notif}'), self.main_loop)
-                    # self.main_loop.call_soon_threadsafe(self.send_telegram_notif, 'asdasdasdasdas')
-                    # self.main_loop.run_until_complete(self.send_telegram_notif(f'UPS event: {notif}'))
 
             except Exception as e:
                 logger.warning(f'Exception in processing server fifo: {e}', exc_info=e)
